Extract PhotonActionSheetSiteHeaderView into it’s own file
